Move (ui-common): const assertion 적용 상수 `constants.ts` 파일로 이동
