+++ packages/ui/common/src/constants.ts
@@ -1,0 +1,7 @@
+export const ELEMENT_SIZES = ['xs', 'sm', 'md', 'lg', 'xl'] as const;
+export const ELEMENT_SPACINGS = [0, 2, 4, 6, 8, 10, 12, 16, 20] as const;
+export const ELEMENT_STATUS = [
+  'isDisabled',
+  'isSelected',
+  'isLoading',
+] as const;
