+++ packages/ui/common/src/types.ts
@@ -1,12 +1,6 @@
 import type { Globals } from 'csstype';
 
-export const ELEMENT_SIZES = ['xs', 'sm', 'md', 'lg', 'xl'] as const;
-export const ELEMENT_SPACINGS = [0, 2, 4, 6, 8, 10, 12, 16, 20] as const;
-export const ELEMENT_STATUS = [
-  'isDisabled',
-  'isSelected',
-  'isLoading',
-] as const;
+import { ELEMENT_SIZES, ELEMENT_SPACINGS, ELEMENT_STATUS } from './constants';
 
 // Element
 export type ElementSize = (typeof ELEMENT_SIZES)[number];
